ViewControllerの処理のうちViewの直接の変更を伴う処理と状態だけを変更する処理に分割する。
一時的な移行処置としてGameManagerクラスを導入し状態だけを変更する処理を移譲する。
ViewControllerからGameManagerの状態変更を監視し、ViewControllerによるViewの変更は状態変更の通知を受けて行う。
これにより、
1. GameManagerの状態の変更
2. ViewControllerによるViewの更新
3. ユーザーのアクション
4. 1に戻る

のように
アクション=>状態の変更=>ビューの変更
データの流れが一方通行になりプログラムの見通しが改善する
