+++ Reversi/Canceller.swift
@@ -1,0 +1,16 @@
+import Foundation
+
+final class Canceller {
+    private(set) var isCancelled: Bool = false
+    private let body: (() -> Void)?
+
+    init(_ body: (() -> Void)?) {
+        self.body = body
+    }
+
+    func cancel() {
+        if isCancelled { return }
+        isCancelled = true
+        body?()
+    }
+}
